Button and image changes with message
I attempted to make the button the same button, but switches words for each click, from Try it! to Done to Play it Back to Next word. and also wanted to make the Javascript switch the images, so we can add the images on the app.js file. style.css doesn't work compared to base.html's css